requirements.txt
fastapi[all]
requests
uvicorn
gunicorn
sqlalchemy
pydantic
psycopg2
python-dotenv